requirements.txt
torch
tqdm
transformers
datasets
pandas
numpy
wandb
ftfy
regex
scikit-learn
click
joblib
pyyaml
nltk
hydra-core
pytorch-lightning
hydra_colorlog
rich
six
filelock